:art: Move other helper scripts to dev
