+++ cellxgene_lamin/__init__.py
@@ -16,7 +16,6 @@ This is the complete API reference:
 
 __version__ = "0.2.6"  # denote a pre-release for 0.1.0 with 0.1rc1
 
-from . import datasets
-from ._curate import Curate
-from ._cxg_rest import get_collections_from_cxg, get_datasets_from_cxg
-from ._fields import CellxGeneFields
+from .curate import Curate
+from .dev import datasets
+from .fields import CellxGeneFields
